respecting draft inside of getCachedTopLevelNavItems and moving donateNavItem inside cached fn
