+++ src/components/Header/Header.tsx
@@ -7,7 +7,7 @@ import invariant from 'tiny-invariant'
 import { ImageMedia } from '../Media/ImageMedia'
 import { DesktopNav } from './DesktopNav.client'
 import { MobileNav } from './MobileNav.client'
-import { convertToNavLink, getCachedTopLevelNavItems, TopLevelNavItem } from './utils'
+import { getCachedTopLevelNavItems } from './utils'
 
 export async function Header({ center }: { center: string }) {
   const { isEnabled: draft } = await draftMode()
@@ -41,38 +41,7 @@ export async function Header({ center }: { center: string }) {
     `Depth not set correctly when querying settings. USFS Logo for tenant ${center} exists but is not an object.`,
   )
 
-  const navigationRes = await payload.find({
-    collection: 'navigations',
-    depth: 99,
-    draft,
-    where: {
-      'tenant.slug': {
-        equals: center,
-      },
-    },
-  })
-
-  const navigation = navigationRes.docs[0]
-
-  if (!navigation) {
-    payload.logger.error(`Navigation for tenant ${center} missing`)
-    return <></>
-  }
-
-  const topLevelNavItems = await getCachedTopLevelNavItems(center)(center)
-
-  let donateNavItem: TopLevelNavItem | undefined = undefined
-
-  if (navigation.donate?.link) {
-    const link = convertToNavLink(navigation.donate.link)
-
-    if (link) {
-      donateNavItem = {
-        label: link.label,
-        link,
-      }
-    }
-  }
+  const { topLevelNavItems, donateNavItem } = await getCachedTopLevelNavItems(center, draft)()
 
   return (
     <header className="bg-header lg:shadow-sm lg:border-b">
