+++ src/components/Header/utils.ts
@@ -233,7 +233,7 @@ export const getTopLevelNavItems = async ({
   navigation: Navigation
   avalancheCenterMetadata?: AvalancheCenter
   avalancheCenterPlatforms: AvalancheCenterPlatforms
-}): Promise<TopLevelNavItem[]> => {
+}): Promise<{ topLevelNavItems: TopLevelNavItem[]; donateNavItem?: TopLevelNavItem }> => {
   let forecastsNavItem: TopLevelNavItem = {
     link: {
       label: 'Forecasts',
@@ -362,13 +362,14 @@ export const getTopLevelNavItems = async ({
     weatherNavItem.items.splice(weatherStationsNavItemIndex, 1)
   }
 
-  return [
+  const topLevelNavItems: TopLevelNavItem[] = [
     ...(avalancheCenterPlatforms.forecasts ? [forecastsNavItem] : []),
     weatherNavItem,
     ...(avalancheCenterPlatforms.obs ? [observationsNavItem] : []),
     ...topLevelNavItem({ tab: navigation.education, label: 'Education' }),
     ...topLevelNavItem({ tab: navigation.accidents, label: 'Accidents' }),
     {
+      label: 'Blog',
       link: {
         label: 'Blog',
         type: 'internal',
@@ -376,6 +377,7 @@ export const getTopLevelNavItems = async ({
       },
     },
     {
+      label: 'Events',
       link: {
         label: 'Events',
         type: 'internal',
@@ -385,16 +387,32 @@ export const getTopLevelNavItems = async ({
     ...topLevelNavItem({ tab: navigation.about, label: 'About' }),
     ...topLevelNavItem({ tab: navigation.support, label: 'Support' }),
   ]
+
+  let donateNavItem: TopLevelNavItem | undefined = undefined
+
+  if (navigation.donate?.link) {
+    const link = convertToNavLink(navigation.donate.link)
+
+    if (link) {
+      donateNavItem = {
+        label: link.label,
+        link,
+      }
+    }
+  }
+
+  return { topLevelNavItems, donateNavItem }
 }
 
-export const getCachedTopLevelNavItems = (center: string) =>
+export const getCachedTopLevelNavItems = (center: string, draft: boolean = false) =>
   unstable_cache(
-    async (center: string): Promise<TopLevelNavItem[]> => {
+    async (): Promise<{ topLevelNavItems: TopLevelNavItem[]; donateNavItem?: TopLevelNavItem }> => {
       const payload = await getPayload({ config: configPromise })
 
       const navigationRes = await payload.find({
         collection: 'navigations',
         depth: 99,
+        draft,
         where: {
           'tenant.slug': {
             equals: center,
@@ -406,7 +424,7 @@ export const getCachedTopLevelNavItems = (center: string) =>
 
       if (!navigation) {
         payload.logger.error(`Navigation for tenant ${center} missing`)
-        return []
+        return { topLevelNavItems: [] }
       }
 
       const avalancheCenterMetadata = await getAvalancheCenterMetadata(center)
@@ -426,7 +444,7 @@ export const getCachedTopLevelNavItems = (center: string) =>
 
 export async function getCanonicalUrlForSlug(center: string, slug: string): Promise<string | null> {
   try {
-    const topLevelNavItems = await getCachedTopLevelNavItems(center)(center)
+    const { topLevelNavItems } = await getCachedTopLevelNavItems(center)()
     const navigationItem = findNavigationItemBySlug(topLevelNavItems, slug)
 
     if (navigationItem?.link?.type === 'internal') {
